requirements.txt

streamlit==1.20.0
tensorflow>=2.10.0,<2.16.0
keras>=2.10.0,<2.16.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
requests>=2.25.0
